+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.1.12] - 19-03-2024
+### Changed
+- Fix problem with `stream=True`
+- Remove `format` parameter: Now it will be always `text`. 
+
 ## [0.1.11] - 01-03-2024
 ### Changed
 - Remove unused packages